commit 13/02/2025

+++ src/main/java/com/thean/dreamshops/model/Product.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Table(name = "tbl_product")
@@ -23,7 +24,7 @@ public class Product {
 
     private String brand;
 
-    private double price;
+    private BigDecimal price;
 
     private int inventory;
 
@@ -36,7 +37,7 @@ public class Product {
     @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Image> images;
 
-    public Product(String name, String brand, double price, int inventory, String description, Category category) {
+    public Product(String name, String brand, BigDecimal price, int inventory, String description, Category category) {
         this.name = name;
         this.brand = brand;
         this.price = price;
